+++ CHANGES.rst
@@ -1,8 +1,9 @@
 1.0.3 (unreleased)
 ==================
 
-- Nothing changed yet.
+- Add web plugin to allow to POST messages to channel via http
 
+- Drop py33 support
 
 1.0.2 (2017-11-13)
 ==================